:wrench: Fix/currency int to double issue

+++ lib/features/cashier/presentation/widgets/drawer/explore_product_drawer_tablet.dart
@@ -82,7 +82,7 @@ class _ExploreProductDrawerTabletState
                       builder: (context, state) => switch (state) {
                         CartDetailLoadSuccess(:final previewOrderPrice) =>
                           SelectPaymentMethodTablet(
-                            amount: int.parse(previewOrderPrice.total),
+                            amount: double.parse(previewOrderPrice.total),
                             onPaymentCash: _onCashPaid,
                           ),
                         CartDetailLoadFailure(:final error) => Center(
